Added support for native go tooling
restuctured the directories and verified the generate functions.
added a replace statement in the go.mod file to resolve generated imports
moved the unit tests back to pkg/protos, even though the same package existed, the go could not resolve the tests.
bazel updates to BUILD to support new structure.